Add oauth2 consent handler
